+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Process uploads with `application/ogg` media type as `audio/ogg`.
 - Strip query parameters when canonicalizing 'ap' URIs.
 - Verify ownership of embedded objects submitted by FEP-ae97 clients.
+- Allow empty `status` field in `/api/v1/statuses` form data.
 
 ### Fixed